HunterMatrix/ai-security/requirements.txt
# AI安全平台依赖包

# Web框架
aiohttp==3.8.5
aiohttp-cors==0.7.0
websockets==11.0.3

# 机器学习
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
joblib==1.3.2

# 深度学习
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2
sentence-transformers==2.2.2

# 自然语言处理
spacy==3.6.1
nltk==3.8.1
openai==0.28.0

# 数据处理
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# 安全分析
yara-python==4.3.0
pefile==2023.2.7
python-magic==0.4.27
ssdeep==3.4

# 网络分析
scapy==2.5.0
dpkt==1.9.8
pyshark==0.6

# 系统监控
psutil==5.9.5
watchdog==3.0.0

# 工具库
requests==2.31.0
pyyaml==6.0.1
python-dateutil==2.8.2
tqdm==4.66.1

# 开发工具
pytest==7.4.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1
